requirements.txt
Django==3.2
msal==1.10.0
python_dateutil==2.8.2
PyYAML==6.0
requests==2.27.1
